requirements.txt
streamlit
sqlalchemy
pillow
numpy
pandas
folium
streamlit-folium
